requirements.txt
# Django
# ------------------------------------------------------------------------------
django==4.2 # https://www.djangoproject.com/
djangorestframework==3.15.1 # https://github.com/encode/django-rest-framework
django-extensions==3.2.3  # https://github.com/django-extensions/django-extensions
django-cors-headers==4.3.1  # https://github.com/adamchainz/django-cors-headers
dj-rest-auth==5.1.0 # https://dj-rest-auth.readthedocs.io/en/latest/
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
django-filter==24.2
ipython==8.23.0
gunicorn==22.0.0
drf-spectacular==0.27.2
django-storages==1.14.1
boto3==1.34.93
pillow==10.3.0
django-phonenumber-field==7.3.0
phonenumberslite==8.13.36

# Database
# ------------------------------------------------------------------------------
psycopg==3.1.18
psycopg-binary==3.1.18

# Testing
# ------------------------------------------------------------------------------
pytest==8.1.1  # https://github.com/pytest-dev/pytest
pytest-django==4.8.0  # https://github.com/pytest-dev/pytest-django
pytest-cov==5.0.0
pytest-xdist==3.5.0
factory_boy==3.3.0
pytest-mock==3.14.0

# Code quality
# ------------------------------------------------------------------------------
ruff==0.5.0  # https://github.com/astral-sh/ruff
flake8==7.0.0  # https://github.com/PyCQA/flake8
flake8-isort==6.1.1  # https://github.com/gforcada/flake8-isort
coverage==7.4.4  # https://github.com/nedbat/coveragepy
black==24.3.0  # https://github.com/psf/black
pylint-django==2.5.5  # https://github.com/PyCQA/pylint-django
pylint-celery==0.3  # https://github.com/PyCQA/pylint-celery
pre-commit==3.7.0  # https://github.com/pre-commit/pre-commit

# AI
# ------------------------------------------------------------------------------
transformers==4.41.0
torch==2.3.0
accelerate==0.30.1
